removed unnecessary code and files
