+++ Denifia.Stardew.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26228.4
+VisualStudioVersion = 15.0.26228.10
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SendItems", "SendItems\SendItems.csproj", "{D74FB037-CC7C-4047-8384-2C2E37999A7A}"
 EndProject
@@ -16,15 +16,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".metadata", ".metadata", "{
 		readme.md = readme.md
 	EndProjectSection
 EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Pathoschild.Stardew.Common", "Pathoschild.Stardew.Common\Pathoschild.Stardew.Common.shproj", "{EB97DAD7-6B56-40B9-8622-E1A795589D5F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuyRecipes", "BuyRecipes\BuyRecipes.csproj", "{7DACA8E6-D69F-43BB-8442-CDB25F7C863B}"
 EndProject
 Global
-	GlobalSection(SharedMSBuildProjectFiles) = preSolution
-		Pathoschild.Stardew.Common\Pathoschild.Stardew.Common.projitems*{d74fb037-cc7c-4047-8384-2c2e37999a7a}*SharedItemsImports = 4
-		Pathoschild.Stardew.Common\Pathoschild.Stardew.Common.projitems*{eb97dad7-6b56-40b9-8622-e1a795589d5f}*SharedItemsImports = 13
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
