--- Pathoschild.Stardew.Common/Pathoschild.Stardew.Common.projitems
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>eb97dad7-6b56-40b9-8622-e1a795589d5f</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Pathoschild.Stardew.Common</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)CommonConstants.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CommonHelper.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UpdateHelper.cs" />
-  </ItemGroup>
-</Project>
