+++ SendItems/Framework/ModConstants.cs
@@ -1,9 +1,9 @@
-﻿namespace Pathoschild.Stardew.Common
+﻿namespace Denifia.Stardew.SendItems.Framework
 {
-    /// <summary>Constant mod values.</summary>
-    internal class CommonConstants
+    internal class ModConstants
     {
         /// <summary>The URL which contains the latest versions for each mod.</summary>
         public const string UpdateUrl = "https://raw.githubusercontent.com/denifia/StardewMods/master/latest-versions.json";
+        public const string ModName = "Send Items";
     }
 }
